hello_world/requirements.txt
MarkupSafe==1.0
flask-lambda-python36
